hacked restkit to fix a bug in POST payload
I needed to submit the same key more than once
in the payload ( key[]=value rails style ). The
problem is that restkit only accepts dicts or
strings as payload, and strings are http-
escaped. So I copied restkit's escaping
method and modified it so that it would accept
lists. I'm planning on submitting a patch
to restkit.